+++ lib/modules/forms/forms_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: forms
-// Updated at: 2026-05-16 16:23:06
-// Commit: fix(forms): fix memory leak by cancelling forms subscription
+// Updated at: 2026-05-17 09:53:44
+// Commit: docs(forms): update readme and API contracts for forms
 
 class formsComponentState {
-  final String commitRef = '1209';
-  final String lastMessage = 'fix(forms): fix memory leak by cancelling forms subscription';
-  final DateTime timestamp = DateTime.parse('2026-05-16T16:23:06+05:00');
+  final String commitRef = '1219';
+  final String lastMessage = 'docs(forms): update readme and API contracts for forms';
+  final DateTime timestamp = DateTime.parse('2026-05-17T09:53:44+05:00');
 }